refactor(경로): dto 패키지 정리
