+++ src/main/java/inhagdsc/mamasteps/map/dto/requestProfile/EditRequestProfileResponse.java
@@ -1,15 +1,17 @@
-package inhagdsc.mamasteps.map.dto;
+package inhagdsc.mamasteps.map.dto.requestProfile;
 
 import inhagdsc.mamasteps.map.domain.LatLng;
 import inhagdsc.mamasteps.map.domain.RouteRequestProfileEntity;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import java.util.List;
 
 @Getter
 @Setter
-public class GetRequestProfileResponse {
+@NoArgsConstructor
+public class EditRequestProfileResponse {
     private int targetTime;
     private double walkSpeed;
     private LatLng origin;
@@ -18,7 +20,7 @@ public class GetRequestProfileResponse {
     private String createdAt;
     private String updatedAt;
 
-    public GetRequestProfileResponse(RouteRequestProfileEntity routeRequestProfileEntity) {
+    public EditRequestProfileResponse(RouteRequestProfileEntity routeRequestProfileEntity) {
         this.targetTime = routeRequestProfileEntity.getTargetTime();
         this.walkSpeed = routeRequestProfileEntity.getWalkSpeed();
         this.origin = routeRequestProfileEntity.getOrigin();
@@ -26,5 +28,21 @@ public class GetRequestProfileResponse {
         this.endCloseWaypoints = routeRequestProfileEntity.getEndCloseWaypoints();
         this.createdAt = routeRequestProfileEntity.getCreatedAt();
         this.updatedAt = routeRequestProfileEntity.getUpdatedAt();
+    }
+
+    public List<LatLng> getStartCloseWaypoints() {
+        return startCloseWaypoints;
+    }
+
+    public void setStartCloseWaypoints(List<LatLng> startCloseWaypoints) {
+        this.startCloseWaypoints = startCloseWaypoints;
+    }
+
+    public List<LatLng> getEndCloseWaypoints() {
+        return endCloseWaypoints;
+    }
+
+    public void setEndCloseWaypoints(List<LatLng> endCloseWaypoints) {
+        this.endCloseWaypoints = endCloseWaypoints;
     }
 }
